+++ src/TableView.tsx
@@ -79,11 +79,11 @@ export default defineComponent({
       required: false,
       default: null
     },
-    menu: { // 单独制定分页器数据，不从listData中取值
+    menu: {
       type: Object as PropType<TableViewProps['menu']>,
       required: false
     },
-    tableProps: { // 单独制定分页器数据，不从listData中取值
+    tableProps: {
       type: Object as PropType<TableViewProps['tableProps']>,
       required: false
     },